remove tests for old behavior
I removed the tests for importing definitions that get `import`ed in the .proto files because this behavior is being replaced by the ability to import multiple proto files.
I'm also removing one of the tests for the `inject` flag since its behavior has now changed. Inject now indicates that you want to generate modules with the namespacing from the proto files not not nest those under the namespace of the module that calls `use Protobuf`